Button:PressingMilliseconds: memory optimization (uint32_t -> uint16_t)

+++ lib/DnWiFiDoorLock/DnWiFiDoorLock/App.h
@@ -84,7 +84,7 @@ namespace DnWiFiDoorLock {
 
                 const uint8_t notPressingAngle;
 
-                const uint32_t pressingMilliseconds;
+                const uint16_t pressingMilliseconds;
             } const furnaceHeaterButton;
         };
